# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "DocServer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
        "/auth/forgot-password": {
            "post": {
                "description": "Initiates the password reset process by requesting a One-Time Password (OTP) to be sent (conceptually) to the user's registered email address.\n\nProvide the `email` address associated with the account you want to reset the password for.\n**Security Note:** To prevent attackers from figuring out which emails are registered (\"email enumeration\"), this endpoint will *always* return a `202 Accepted` response, regardless of whether the email exists in the system or not.\nIf the email *does* exist, the server generates an OTP, stores it temporarily, and (in a real system) would send it via email. The OTP is needed for the `/auth/reset-password` step.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Request Password Reset Code (OTP)",
                "parameters": [
                    {
                        "description": "The email address for the account needing a password reset.",
                        "name": "forgotPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Request Accepted. If the email address is registered, an OTP has been generated (and would typically be emailed). Check your email for the code."
                    },
                    "400": {
                        "description": "Bad Request: The request body is invalid (e.g., missing email or invalid format).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while processing the request (e.g., OTP generation failed).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticates a user using their registered email and password.\n\nIf the credentials are correct, the server generates a JSON Web Token (JWT). This token acts like a temporary key or session ID.\nYou need to include this JWT in the `Authorization` header (as a Bearer token) for subsequent requests to protected endpoints (like accessing your profile or documents).\nExample Header: `Authorization: Bearer \u003cyour_token_here\u003e`",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Log In to Your Account",
                "parameters": [
                    {
                        "description": "Your email and password.",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Successful. The response body contains the JWT access token.",
                        "schema": {
                            "$ref": "#/definitions/api.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request: The data you sent is invalid (e.g., missing email or password, incorrect JSON format).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: The email or password you provided is incorrect. Please check your credentials.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server during login (e.g., database issue, error generating the JWT).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Indicates the intention to log out. Since JWTs are stateless (the server doesn't keep track of active tokens), true logout happens on the client-side.\n\n**Action Required by Client:** To effectively log out, the client application (e.g., your web browser or mobile app) MUST delete or discard the stored JWT access token.\nCalling this endpoint doesn't invalidate the token on the server, but it serves as a conventional way to signal the end of a session in API design.",
                "tags": [
                    "Authentication"
                ],
                "summary": "Log Out (Client-Side Action)",
                "responses": {
                    "204": {
                        "description": "Logout Signaled. No content is returned. Remember to discard the JWT on the client."
                    },
                    "401": {
                        "description": "Unauthorized: Although logout is client-side, this endpoint might still require a valid token to be called as per API design consistency.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "description": "Completes the password reset process by setting a new password for the account.\n\nYou must provide:\n*   The `email` address of the account.\n*   The `otp` (One-Time Password) received after calling `/auth/forgot-password`.\n*   The desired `new_password` (must meet minimum length requirements, e.g., 8 characters).\n\nThe server will first verify if the provided OTP is correct and hasn't expired for the given email. If valid, it will hash the `new_password` and update the user's account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Set New Password Using Reset Code (OTP)",
                "parameters": [
                    {
                        "description": "Email, OTP, and the new password.",
                        "name": "resetPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Password Reset Successful. Your new password is now active. You can log in using it. No content is returned in the response body."
                    },
                    "400": {
                        "description": "Bad Request: The request body is invalid (e.g., missing fields, new password too short).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: The provided OTP is incorrect, expired, or does not match the email address.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found: The profile associated with the email address could not be found (e.g., it might have been deleted after the OTP was requested).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server (e.g., hashing the new password failed, database update failed).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Creates a new user profile in the system. This is the first step for a new user to start using the service.\n\nYou need to provide your desired `email`, a secure `password` (minimum 8 characters), your `first_name`, and `last_name`.\nThe server will securely hash the password before storing it (meaning the original password is never saved directly).\nIf the email address is already registered, the request will fail.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a New User Account",
                "parameters": [
                    {
                        "description": "User registration details. All fields except 'extra' are required.",
                        "name": "signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account Created Successfully. The response body contains the details of the newly created profile (excluding the password hash).",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request: The data you sent is invalid (e.g., missing required fields, invalid email format, password too short) OR the email address is already in use by another account.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while creating the account (e.g., password hashing failed, database connection issue).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "api.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "api.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "api.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "email",
                "new_password",
                "otp"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string",
                    "minLength": 8
                },
                "otp": {
                    "type": "string"
                }
            }
        },
...
        "api.SignupRequest": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "extra": {},
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "description": "Add basic password length validation",
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "models.Profile": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "description": "UTC",
                    "type": "string"
                },
                "email": {
                    "description": "Unique, used for login",
                    "type": "string"
                },
                "extra": {
                    "description": "User-defined data"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID (UUID, dashless)",
                    "type": "string"
                },
                "last_modified_date": {
                    "description": "UTC",
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "utils.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}